refactor: wrap utils to folder
